* app/WorldEntity.h, visual/Isometric.cpp, visual/Isometric.h: 	  Fix movement prediction so it works in fracitons of a second.
	* app/GameClient.cpp, app/GameClient.h: Add movement prediction to
	  the camera control.

	* visual/Isometric.cpp, visual/DemeterScene.cpp: Fix up media
	  pathnames.

Al Riddoch  <ajr@ecs.soton.ac.uk>

+++ app/WorldEntity.h
@@ -12,7 +12,7 @@
 
 class WorldEntity : public Eris::Entity
 {
-    unsigned int updateTime;
+    float updateTime;
   public:
     VisualEntity & visEntity;
 
@@ -20,7 +20,7 @@ class WorldEntity : public Eris::Entity
 
     WorldEntity(const Atlas::Objects::Entity::GameEntity &ge);
 
-    const unsigned int getTime() {
+    const float getTime() {
         return updateTime;
     }
 };
